+++ lib/pidSupport.py
@@ -151,7 +151,7 @@ class PidSupport:
             raise RuntimeError("Corrupted lock file: no PID")
 
         try:
-            pid = long(pidarr[1])
+            pid = int(pidarr[1])
         except:
             raise RuntimeError("Corrupted lock file: invalid PID")
         
@@ -203,7 +203,7 @@ class PidWParentSupport(PidSupport):
             raise RuntimeError("Corrupted lock file: no PID")
 
         try:
-            pid = long(pidarr[1])
+            pid = int(pidarr[1])
         except:
             raise RuntimeError("Corrupted lock file: invalid PID")
         
@@ -212,7 +212,7 @@ class PidWParentSupport(PidSupport):
             raise RuntimeError("Corrupted lock file: no Parent PID")
 
         try:
-            parent_pid = long(pidarr[1])
+            parent_pid = int(pidarr[1])
         except:
             raise RuntimeError("Corrupted lock file: invalid Parent PID")
